apps/compliance_api/requirements.txt
# Web Framework & Server
fastapi
uvicorn
python-dotenv
python-multipart

# AI & LangChain
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pypdf
python-docx
tiktoken

# Data & Database
sqlalchemy
pydantic
pandas

# Frontend (Dashboard)
streamlit
requests
pillow